Problem: reading large images (in my case, it was a 40792x20480 RGBA PNG one) causes a crash, when the product row-number * width-step overflows a 32-bit integer
Solution: change the type of the width step to be `size_t` (and not `long`)

+++ dlib/python/numpy_image.h
@@ -319,7 +319,7 @@ namespace dlib
     }
 
     template <typename pixel_type>
-    long width_step (const numpy_image<pixel_type>& img)
+    size_t width_step (const numpy_image<pixel_type>& img)
     {
         if (img.size()==0)
             return 0;
